BundleFactory: Add function to load codec bundles
Codec bundles are structurally the same as search bundles, but perform
a logically different function: they specify a set of filters to perform
some initial processing on each object (e.g. RGB decoding and thumbnail
generation).  Because they must be presented via a different mechanism in
the search UI, they are given a different filename extension.