Remove search ID andat data from object
In the past, search ID was used for relating objects to searches because
searches for different channels were done on different channels. Now all
searches are done on the same channel, and it is unnecessary to relate
objects to searches. Therefore, remove search ID from objects.

Also, remove data field from object since it is not used anymore.
Object data is now stored in the attributes list with the zero-length
string key.

Change get_object RPC to 2 (from 1).