diamondd: Change statistics reply to use array of stats
Currently, the statistics RPC uses a static list of hardcoded
statistics values. This becomes an inconvenience when the list has to be
expanded in the case different types of statistics are added. By having
the list of statistics values as an array, the list of statistics is more
easily expandable.

Change statistics RPC number (from 15 to 16).