diamondd: Remove attribute/blob names and array lengths limits
diamondd used to be implemented in C. As it is inconvenient to manage
variable length strings and arrays, there used to be length limits for
attribute/blob names and for arrays. diamondd is now implemented in
Python, and the limits are not necessary anymore as it is easy to
manage variable length strings and arrays.