opendiamond-java: Change ServerStatistics to use map
ServerStatistics.java uses static list of harcoded statistics values.
This becomes inconvenient when the list has to be changed. By having
the list of statistics values as a key-value map, the list is more
easily modifiable. Add a new API to return a statistics value using a
string key.

Let ServerStatistics.java to have three static string constants and
three getters to return objects total, objects processed and objects
dropped. This is because current Diamond applications do not support
the new API that uses the string key.

Also change the lists and maps that are directly available to users to
be unmodifiable.

Fix statistics RPC number to the next available one (29).