opendiamond-java: Fix XDR_reexecute to use DataOutputStream
Using ByteBuffer to manipulate byte arrays is bad. It is much more clean
and concise to use DataOutputStream and ByteArrayOutputStream. Fix
XDR_reexecute to use the latter two to generate the protocol message.

Also, check for null or an empty attributes lists so that in those cases
no attributes list is included in the request message.

Fix the reexecute RPC number to the next available one (30).