opendiamond-java: Change ServerStatistics class to use map
ServerStatistics class currently has an API to return only specific
statistics values. The statistics values are stored in a map, but the
keys are not visible to users, and thus users can only retrieve values
that the API methods return. Add new APIs to return the map of global
server statistics and the map of per-filter statistics. Using the new
methods, users can view the entire set of global server and per-filter
statistics.

Add new class FilterStatistics to assist the above explained
functionality.